// src/enums/movie-sort-options.ts
export enum MovieSortOptions {
    PopularityDesc = 'popularity.desc',
    PopularityAsc = 'popularity.asc',
    VoteAverageDesc = 'vote_average.desc',
    VoteAverageAsc = 'vote_average.asc',
    OriginalTitleDesc = 'original_title.desc',
    OriginalTitleAsc = 'original_title.asc',
    ReleaseDateDesc = 'release_date.desc',
    ReleaseDateAsc = 'release_date.asc'
}